+++ src/web/add_object.cc
@@ -43,7 +43,10 @@
 #include "util/tools.h"
 #include "util/xml_to_json.h"
 
-void Web::AddObject::addContainer(int parentID, const std::string& title, const std::string& upnp_class)
+void Web::AddObject::addContainer(
+    int parentID,
+    const std::string& title,
+    const std::string& upnp_class)
 {
     auto cont = content->addContainer(parentID, title, upnp_class);
 
@@ -54,7 +57,11 @@ void Web::AddObject::addContainer(int parentID, const std::string& title, const 
     }
 }
 
-std::shared_ptr<CdsItem> Web::AddObject::addItem(int parentID, const std::string& title, const std::string& upnp_class, const fs::path& location)
+std::shared_ptr<CdsItem> Web::AddObject::addItem(
+    int parentID,
+    const std::string& title,
+    const std::string& upnp_class,
+    const fs::path& location)
 {
     auto item = std::make_shared<CdsItem>();
     item->setParentID(parentID);
@@ -86,7 +93,12 @@ std::shared_ptr<CdsItem> Web::AddObject::addItem(int parentID, const std::string
     return item;
 }
 
-std::shared_ptr<CdsItemExternalURL> Web::AddObject::addUrl(int parentID, const std::string& title, const std::string& upnp_class, bool addProtocol, const fs::path& location)
+std::shared_ptr<CdsItemExternalURL> Web::AddObject::addUrl(
+    int parentID,
+    const std::string& title,
+    const std::string& upnp_class,
+    bool addProtocol,
+    const fs::path& location)
 {
     auto item = std::make_shared<CdsItemExternalURL>();
     std::string protocolInfo;
@@ -102,27 +114,31 @@ std::shared_ptr<CdsItemExternalURL> Web::AddObject::addUrl(int parentID, const s
         item->addMetaData(MetadataFields::M_DESCRIPTION, desc);
     }
 
-    std::string mime = param("mime-type");
-    if (mime.empty())
-        mime = MIMETYPE_DEFAULT;
-    item->setMimeType(mime);
-
     std::string flags = param("flags");
     if (!flags.empty())
         item->setFlags(CdsObject::makeFlag(flags));
 
-    if (addProtocol) {
-        std::string protocol = param("protocol");
-        if (!protocol.empty())
-            protocolInfo = renderProtocolInfo(mime, protocol);
-        else
-            protocolInfo = renderProtocolInfo(mime);
-    } else
-        protocolInfo = renderProtocolInfo(mime);
+    {
+        std::string mime = param("mime-type");
+        if (mime.empty())
+            mime = MIMETYPE_DEFAULT;
+        item->setMimeType(mime);
 
-    auto resource = std::make_shared<CdsResource>(ContentHandler::DEFAULT, ResourcePurpose::Content);
-    resource->addAttribute(ResourceAttribute::PROTOCOLINFO, protocolInfo);
-    item->addResource(resource);
+        if (addProtocol) {
+            std::string protocol = param("protocol");
+            if (!protocol.empty())
+                protocolInfo = renderProtocolInfo(mime, protocol);
+            else
+                protocolInfo = renderProtocolInfo(mime);
+        } else
+            protocolInfo = renderProtocolInfo(mime);
+
+        auto resource = std::make_shared<CdsResource>(ContentHandler::DEFAULT, ResourcePurpose::Content);
+        resource->addAttribute(ResourceAttribute::PROTOCOLINFO, protocolInfo);
+        item->addResource(resource);
+    }
+
+    readResources(item);
 
     return item;
 }