+++ web/js/jquery.gerbera.editor.js
@@ -46,6 +46,10 @@
     video: { oclass: 'object.item.videoItem', type: 'item' },
     videoBroadcast: { oclass: 'object.item.videoItem.videoBroadcast', type: 'external_url', protocol: 'http-get' },
   };
+  const purposeValues = [
+    { label: 'Thumbnail', type: 'thumbnail' },
+    { label: 'Subtitle', type: 'subtitle' },
+  ];
   const objectFlags = [
     "Restricted",
     "Searchable",
@@ -56,6 +60,7 @@
     "OnlineService",
     "OggTheora",
   ];
+  let resources = [];
   function addNewItem(modal, itemData) {
     const itemType = itemData.type;
     const item = itemData.item;
@@ -75,11 +80,15 @@
       editFlagBox[flag] = modal.find('#editFlag-' + flag);
     });
 
-    reset(modal);
 
     editObjectType.off('click').on('click', function () {
       const editObjectType = $(this);
-      addNewItem(modal, { type: editObjectType.val(), item: item, onSave: itemData.onSave })
+      addNewItem(modal, {
+        type: editObjectType.val(),
+        item: item,
+        onSave: itemData.onSave,
+        onData: itemData.onData
+      })
     });
 
     modal.find('#objectIdGroup').hide();
@@ -103,6 +112,8 @@
     hideFields(hiddenFields);
 
     editClass.val(defaultClasses[itemType].oclass);
+    modal.find('.modal-dialog').removeClass('modal-xl');
+    $("#editCol").hide();
     if (defaultClasses[itemType].type === 'container') {
       showFields([editObjectType, editTitle, editClass, editFlags, editFlagBox['Searchable']]);
       hideFields([editLocation, editDesc, editMime, editProtocol]);
@@ -114,10 +125,104 @@
       showFields([editObjectType, editTitle, editLocation, editClass, editDesc, editMime, editProtocol, editFlags,
         editFlagBox['ProxyUrl'],
         /* editFlagBox['OnlineService'] */]);
+      modal.find('.modal-dialog').addClass('modal-xl');
+      $("#editCol").show();
+      const restable = modal.find('#resdata');
+      restable.html('');
+      const tfoot = $('<tfoot></tfoot>');
+      const trow = $('<tr></tr>');
+      const th1 = $('<th></th>');
+      const th2 = $('<th colspan="2"></th>');
+      const addRes = $('<button>Add Resource</button>');
+      addRes.off('click').on('click', function () {
+        resources.push({
+          index: resources.length + 1,
+          changed: ['purpose', 'location', 'mime-type', 'protocol'],
+          purpose: ['thumbnail'],
+          'location': [''],
+          'mime-type': ['image/jpg'],
+          protocol: ['http-get']
+        });
+        addNewItem(modal, {
+          type: itemData.type,
+          item: item,
+          onSave: itemData.onSave,
+          onData: itemData.onData
+        });
+      });
+      addRes.appendTo(th2);
+      th1.appendTo(trow);
+      th2.appendTo(trow);
+      trow.appendTo(tfoot);
+
+      resources.forEach((res, i) => {
+        const tbody = makeDetailsHead(restable, 'Resource', `#${i + 1}`, itemData.onData);
+        makePurposeSelect(tbody, 'purpose', res, purposeValues);
+        makeResourceInput(tbody, ['location', 'mime-type', 'protocol'], res);
+        tbody.appendTo(restable);
+      });
+      tfoot.addClass('itemDetailFoot');
+      tfoot.appendTo(restable);
+      restable.show();
     }
   }
 
+  function makeDetailsHead(table, caption, value, eventHandler) {
+    const thead = $(`<thead><tr><th>${caption}</th><th colspan="2">${value}</th></tr></thead>`);
+    thead.addClass('itemDetailHead');
+    thead.appendTo(table);
+    const tbody = $('<tbody></tbody>');
+
+    if (eventHandler)
+      thead.click(tbody, eventHandler);
+    tbody.appendTo(table);
+    return tbody;
+  }
+
+  function makePurposeSelect(tbody, p, resource, values) {
+    const purposeSelect = $(`<select name="${p}Select" id="${p}Select"></select>`);
+    purposeSelect.change(
+      function (event) {
+        if (resource.changed)
+          resource.changed.push(p);
+        else
+          resource.changed = [p];
+        if (!resource[p][1])
+          resource[p][1] = resource[p][0];
+        resource[p][0] = event.target.value;
+        resource[p][1] = event.target.value;
+      });
+    values.forEach((purpose) => {
+      const selected = resource[p][0] === purpose.type ? 'selected="selected"' : '';
+      const purposeOption = $(`<option ${selected} value="${purpose.type}">${purpose.label}</option>`);
+      purposeOption.appendTo(purposeSelect);
+    });
+    const trow = $('<tr></tr>');
+    const td1 = $(`<td class="detail-column"><span>${p}</span></td>`);
+    td1.appendTo(trow);
+    const td2 = $('<td colspan="2"></td>');
+    purposeSelect.appendTo(td2);
+    td2.appendTo(trow);
+    trow.appendTo(tbody);
+  }
+
+  function makeResourceInput(tbody, properties, resource) {
+    properties.forEach((p) => {
+      appendMetaItem(tbody, p, resource[p][0], null, null,
+        function (event) {
+          if (resource.changed)
+            resource.changed.push(p);
+          else
+            resource.changed = [p];
+          if (!resource[p][1])
+            resource[p][1] = resource[p][0];
+          resource[p][0] = event.target.value;
+        });
+    });
+  }
+
   function reset(modal) {
+    resources = [];
     modal.find('#editObjectType').val('item');
     modal.find('#objectId').val('');
     modal.find('#addParentId').val('');
@@ -148,7 +253,7 @@
     modal.find('#hidebutton').hide();
   }
 
-  function appendMetaItem(tbody, name, value, special = null, rawvalue = null) {
+  function appendMetaItem(tbody, name, value, special = null, rawvalue = null, eventHandler = null) {
     let content, text;
     const nl = '\n';
 
@@ -157,28 +262,33 @@
     content.addClass('detail-column');
     text = $('<span></span>');
     text.text(name).appendTo(content);
-    row.append(content);
+    content.appendTo(row);
     if (rawvalue)
       content = $('<td></td>');
     else
       content = $('<td colspan="2"></td>');
 
-    if (value) {
+    if (eventHandler) {
+      const input = $(`<input class="form-control" value="${value}"/>`);
+      input.change(eventHandler);
+      input.appendTo(content);
+    }
+    else if (value) {
       text = $('<span>' + value.replace(RegExp(nl, 'g'), '<br>') + '</span>');
       text.appendTo(content);
     }
     if (special) {
       special.appendTo(content);
     }
-    row.append(content);
+    content.appendTo(row);
 
     if (rawvalue) {
       content = $('<td></td>');
       text = $('<span>' + rawvalue.replace(RegExp(nl, 'g'), '<br>') + '</span>');
       text.appendTo(content);
-      row.append(content);
+      content.appendTo(row);
     }
-    tbody.append(row);
+    row.appendTo(tbody);
   }
 
   function showDetails(modal) {
@@ -206,10 +316,23 @@
     return result.join('|');
   }
 
+  function readResources(item) {
+    const data = [];
+    resources.forEach((res) => {
+      if (res.changed) {
+        res.changed.forEach((p) => {
+          const output = `resource.${res.index}.${p}`;
+          item[output] = encodeURIComponent(res[p][0]);
+          data.push(output)
+        })
+      }
+    });
+    item.resources = encodeURIComponent(data.join('|'));
+  }
+
   function loadItem(modal, itemData) {
     const item = itemData.item;
     if (item) {
-      reset(modal);
       if (item.image) {
         modal.find('#mediaimage').prop('src', item.image.value);
         modal.find('#mediaimage').show();
@@ -230,27 +353,17 @@
 
       const metatable = modal.find('#metadata');
       const detailButton = modal.find('#detailbutton');
-      let tbody;
       if (item.metadata && item.metadata.metadata.length) {
         detailButton.show();
-        const thead = $('<thead><tr><th colspan="3">Metadata</th></tr></thead>');
-        thead.addClass('itemDetailHead');
-        thead.appendTo(metatable);
-        tbody = $('<tbody></tbody>');
+        const tbody = makeDetailsHead(metatable, 'Metadata', '', itemData.onData);
         for (let i = 0; i < item.metadata.metadata.length; i++) {
           appendMetaItem(tbody, item.metadata.metadata[i].metaname, item.metadata.metadata[i].metavalue);
         }
-        thead.click(tbody, itemData.onData);
-        metatable.append(tbody);
       }
       if (item.flags && item.flags.value) {
-        const thead = $('<thead><tr><th colspan="3">Extras</th></tr></thead>');
-        thead.addClass('itemDetailHead');
-        thead.appendTo(metatable);
-        tbody = $('<tbody></tbody>');
+        const tbody = makeDetailsHead(metatable, 'Extras', '', itemData.onData);
         appendMetaItem(tbody, "flags", item.flags.value);
         tbody.hide();
-        thead.click(tbody, itemData.onData);
         metatable.append(tbody);
       }
 
@@ -258,16 +371,11 @@
       if (item.auxdata && item.auxdata.auxdata.length) {
         detailButton.show();
         auxtable.show();
-        const thead = $('<thead><tr><th colspan="3">Aux Data</th></tr></thead>');
-        thead.addClass('itemDetailHead');
-        thead.appendTo(auxtable);
-        tbody = $('<tbody></tbody>');
+        const tbody = makeDetailsHead(auxtable, 'Aux Data', '', itemData.onData);
         for (let i = 0; i < item.auxdata.auxdata.length; i++) {
           appendMetaItem(tbody, item.auxdata.auxdata[i].auxname, item.auxdata.auxdata[i].auxvalue);
         }
         tbody.hide();
-        thead.click(tbody, itemData.onData);
-        auxtable.append(tbody);
       } else {
         auxtable.hide();
       }
@@ -276,24 +384,44 @@
       if (item.resources && item.resources.resources.length) {
         detailButton.show();
         restable.show();
+        let currentResource = { index: '???', properties: [] };
         for (let i = 0; i < item.resources.resources.length; i++) {
           if (item.resources.resources[i].resname === '----RESOURCE----') {
-            const thead = $(`<thead><tr><th>Resource</th><th colspan="2">#${item.resources.resources[i].resvalue}</th></tr></thead>`);
-            thead.addClass('itemDetailHead');
-            thead.appendTo(restable);
-            tbody = $('<tbody></tbody>');
-            if (i > 0)
-              tbody.hide();
-            thead.click(tbody, itemData.onData);
-            restable.append(tbody);
-          } else if (item.resources.resources[i].resname === 'image') {
-            appendMetaItem(tbody, 'content', null, $('<img width="50px" class="resourceImage" src="' + item.resources.resources[i].resvalue + '"/>'));
-          } else if (item.resources.resources[i].resname === 'link') {
-            appendMetaItem(tbody, 'content', null, $('<a href=' + item.resources.resources[i].resvalue + '>Open</span>'));
+            currentResource = { index: item.resources.resources[i].resvalue, properties: [] };
+            resources.push(currentResource);
           } else {
-            appendMetaItem(tbody, item.resources.resources[i].resname, item.resources.resources[i].resvalue, null, item.resources.resources[i].rawvalue);
+            currentResource.properties.push(item.resources.resources[i].resname);
+            currentResource[item.resources.resources[i].resname] = [item.resources.resources[i].resvalue, item.resources.resources[i].rawvalue];
           }
         }
+        resources.forEach((res, i) => {
+          const tbody = makeDetailsHead(restable, 'Resource', `#${res.index}`, itemData.onData);
+          if (i > 0)
+            tbody.hide();
+          res.properties.forEach((p) => {
+            switch (p) {
+              case 'image': {
+                appendMetaItem(tbody, 'content', null, $(`<img width="50px" class="resourceImage" src="${res[p][0]}"/>`));
+                break;
+              }
+              case 'link': {
+                appendMetaItem(tbody, 'content', null, $(`<a href=${res[p][0]}>Open</span>`));
+                break;
+              }
+              case 'resFile': {
+                if (res.handlerType[0] === 'Exturl')
+                  makeResourceInput(tbody, [p], res);
+                else
+                  appendMetaItem(tbody, p, res[p][0], null, res[p][1]);
+                break;
+              }
+              default: {
+                appendMetaItem(tbody, p, res[p][0], null, res[p][1]);
+                break;
+              }
+            }
+          });
+        });
       } else {
         restable.hide();
       }
@@ -471,6 +599,7 @@
       modal.find('#editFlag-PlaylistRef'),
       modal.find('#editFlag-OggTheora'),
       modal.find('#editFlag-UseResourceRef'),
+      modal.find('#editFlag-OnlineService'),
     ]);
   }
 
@@ -492,6 +621,7 @@
           description: encodeURIComponent(editDesc.val()),
           flags: readFlags(modal),
         };
+        readResources(item);
         break;
       }
       case 'container': {
@@ -500,6 +630,7 @@
           title: encodeURIComponent(editTitle.val()),
           flags: readFlags(modal),
         };
+        readResources(item);
         break;
       }
       case 'external_url': {
@@ -512,6 +643,7 @@
           protocol: editProtocol.val(),
           flags: readFlags(modal),
         };
+        readResources(item);
         break;
       }
     }
@@ -540,6 +672,7 @@
           description: encodeURIComponent(editDesc.val()),
           flags: readFlags(modal),
         };
+        readResources(item);
         break;
       }
       case 'container': {
@@ -550,6 +683,7 @@
           title: encodeURIComponent(editTitle.val()),
           flags: readFlags(modal),
         };
+        readResources(item);
         break;
       }
       case 'external_url': {
@@ -564,6 +698,7 @@
           protocol: editProtocol.val(),
           flags: readFlags(modal),
         };
+        readResources(item);
         break;
       }
     }
@@ -584,6 +719,7 @@
       _super.Constructor.prototype[args.shift()].apply(this, args);
     },
     addNewItem: function (itemData) {
+      reset($(this._element));
       return addNewItem($(this._element), itemData);
     },
     addObject: function () {
@@ -599,6 +735,7 @@
       return hideDetails($(this._element));
     },
     loadItem: function (itemData) {
+      reset($(this._element));
       return loadItem($(this._element), itemData);
     },
     saveItem: function () {